WIP: refactor HTMLViewModel.
