+++ phy/cluster/manual/__init__.py
@@ -7,6 +7,6 @@
 
 from .clustering import Clustering
 from .wizard import Wizard
-from .views import ViewCreator
+from .views import ViewCreator, HTMLViewModel
 from .gui import GUICreator, ClusterManualGUI
 from .session import Session
