+++ phy/plot/view_models/__init__.py
@@ -3,4 +3,4 @@
 
 """View models"""
 
-from .base import BaseViewModel, HTMLViewModel, VispyViewModel
+from .base import BaseViewModel, VispyViewModel
